refactor create accout to avoid using store
