+++ src/vuex/modules/keystore.js
@@ -1,5 +1,5 @@
 import { track } from "scripts/google-analytics"
-
+import { getWallet } from "./wallet"
 export default () => {
   const state = {
     accounts: [],
@@ -35,27 +35,30 @@ export default () => {
       return getSeed()
     },
     async getAddressFromSeed(store, { seedPhrase, network }) {
-      const wallet = await getWallet(seedPhrase, network, store)
+      const networkObject = await this.getNetworkInfo(network, store)
+      const wallet = await getWallet(seedPhrase, networkObject)
       return wallet.cosmosAddress
     },
-    async createKey(
-      { dispatch, state, getters },
-      { seedPhrase, password, name, network }
-    ) {
+    async getNetworkInfo(networkId, store) {
+      return store.getters.networks.find(({ id }) => id === networkId)
+    },
+    async createKey(store, { seedPhrase, password, name, network }) {
       // TODO extract the key storage from the key creation
       const { storeWallet } = await import("@lunie/cosmos-keys")
 
+      // get current network
+      const networkObject = await this.getNetworkInfo(network, store)
       // create a new key pair
-      const wallet = await getWallet(seedPhrase, network, { getters })
+      const wallet = await getWallet(seedPhrase, networkObject)
 
       storeWallet(wallet, name, password, network)
 
-      state.externals.track(`event`, `session`, `create-keypair`)
+      store.state.externals.track(`event`, `session`, `create-keypair`)
 
       // reload accounts as we just added a new one
-      dispatch("loadAccounts")
+      store.dispatch("loadAccounts")
 
-      await dispatch("signIn", {
+      await store.dispatch("signIn", {
         address: wallet.cosmosAddress,
         sessionType: "local"
       })
@@ -69,61 +72,4 @@ export default () => {
     mutations,
     actions
   }
-}
-
-// creates a cosmos addres for the network desired
-function getCosmosAddressCreator(bech32Prefix) {
-  return async seedPhrase => {
-    const { getNewWalletFromSeed } = await import("@lunie/cosmos-keys")
-    return getNewWalletFromSeed(seedPhrase, bech32Prefix)
-  }
-}
-
-// creates a polkadot address
-async function createPolkadotAddress(seedPhrase, addressPrefix) {
-  const [{ Keyring }] = await Promise.all([
-    import("@polkadot/api"),
-    import("@polkadot/util-crypto").then(async ({ cryptoWaitReady }) => {
-      // Wait for the promise to resolve, async WASM or `cryptoWaitReady().then(() => { ... })`
-      await cryptoWaitReady()
-    })
-  ])
-
-  const keyring = new Keyring({
-    ss58Format: Number(addressPrefix),
-    type: "ed25519"
-  })
-  const newPair = keyring.addFromUri(seedPhrase)
-
-  return {
-    cosmosAddress: newPair.address,
-    publicKey: newPair.publicKey,
-    seedPhrase
-  }
-}
-
-async function getWallet(seedPhrase, networkId, store) {
-  const network = await getNetworkInfo(networkId, store)
-  if (!network)
-    throw new Error("Lunie doesn't support address creation for this network.")
-
-  switch (network.network_type) {
-    case "cosmos": {
-      const addressCreator = await getCosmosAddressCreator(
-        network.address_prefix
-      )
-      return await addressCreator(seedPhrase)
-    }
-    case "polkadot": {
-      return await createPolkadotAddress(seedPhrase, network.address_prefix)
-    }
-    default:
-      throw new Error(
-        "Lunie doesn't support address creation for this network."
-      )
-  }
-}
-
-async function getNetworkInfo(networkId, store) {
-  return store.getters.networks.find(({ id }) => id === networkId)
 }
